refactor(tests): split model test files to match source structure
- Split test_jira_models.py (1,952 lines) into 9 focused test files
- Split test_confluence_models.py (822 lines) into 4 focused test files
- Normalize src.mcp_atlassian imports to mcp_atlassian
- Consolidate use_real_jira_data fixture (remove shadowing env-var version)
