+++ tests/unit/models/test_confluence_common_models.py
@@ -1,29 +1,19 @@
 """
-Tests for the Confluence Pydantic models.
+Tests for common Confluence Pydantic models.
 
-These tests validate the conversion of Confluence API responses to structured models
-and the simplified dictionary conversion for API responses.
+Tests for ConfluenceAttachment, ConfluenceUser, ConfluenceSpace,
+ConfluenceVersion, ConfluenceComment, and ConfluenceLabel models.
 """
 
-import pytest
-
-from src.mcp_atlassian.models import (
+from mcp_atlassian.models import (
     ConfluenceAttachment,
     ConfluenceComment,
     ConfluenceLabel,
-    ConfluencePage,
-    ConfluenceSearchResult,
     ConfluenceSpace,
     ConfluenceUser,
     ConfluenceVersion,
 )
-from src.mcp_atlassian.models.constants import EMPTY_STRING
-
-# Optional: Import real API client for optional real-data testing
-try:
-    from src.mcp_atlassian.confluence.client import ConfluenceClient  # noqa: F401
-except ImportError:
-    pass
+from mcp_atlassian.models.constants import EMPTY_STRING
 
 
 class TestConfluenceAttachment:
@@ -466,357 +456,3 @@ class TestConfluenceLabel:
         assert simplified["name"] == "test"
         assert simplified["prefix"] == "my"
         assert simplified["label"] == "test"
-
-
-class TestConfluencePage:
-    """Tests for the ConfluencePage model."""
-
-    def test_from_api_response_with_valid_data(self, confluence_page_data):
-        """Test creating a ConfluencePage from valid API data."""
-        page = ConfluencePage.from_api_response(confluence_page_data)
-
-        assert page.id == "987654321"
-        assert page.title == "Example Meeting Notes"
-        assert page.type == "page"
-        assert page.status == "current"
-
-        # Verify nested objects
-        assert page.space is not None
-        assert page.space.key == "PROJ"
-        assert page.space.name == "Project Space"
-
-        assert page.version is not None
-        assert page.version.number == 1
-        assert page.version.by is not None
-        assert page.version.by.display_name == "Example User (Unlicensed)"
-
-        # Content extraction depends on the implementation
-        # If it's not extracting from the mock data, let's skip this check
-        # assert "<h2>" in page.content
-
-        # Check timestamps
-        assert page.version.when == "2024-01-01T09:00:00.000Z"
-
-    def test_from_api_response_with_empty_data(self):
-        """Test creating a ConfluencePage from empty data."""
-        page = ConfluencePage.from_api_response({})
-
-        # Should use default values
-        assert page.id == "0"
-        assert page.title == ""
-        assert page.type == "page"
-        assert page.status == "current"
-        assert page.space is None
-        assert page.content == ""
-        assert page.content_format == "view"
-        assert page.created == ""
-        assert page.updated == ""
-        assert page.author is None
-        assert page.version is None
-        assert len(page.ancestors) == 0
-        assert isinstance(page.children, dict)
-        assert page.url is None
-
-    def test_from_api_response_with_search_result(self, confluence_search_data):
-        """Test creating a ConfluencePage from search result content."""
-        content_data = confluence_search_data["results"][0]["content"]
-
-        page = ConfluencePage.from_api_response(content_data)
-
-        assert page.id == "123456789"
-        assert page.title == "2024-01-01: Team Progress Meeting 01"
-        assert page.type == "page"
-        assert page.status == "current"
-
-    def test_to_simplified_dict(self, confluence_page_data):
-        """Test converting ConfluencePage to a simplified dictionary."""
-        page = ConfluencePage.from_api_response(confluence_page_data)
-
-        simplified = page.to_simplified_dict()
-
-        assert isinstance(simplified, dict)
-        assert simplified["id"] == "987654321"
-        assert simplified["title"] == "Example Meeting Notes"
-
-        # The keys in the simplified dict depend on the implementation
-        # Let's check for space information in a more flexible way
-        assert page.space is not None
-        assert page.space.key == "PROJ"
-
-        # Check space information - could be a string or a dict
-        if "space_key" in simplified:
-            assert simplified["space_key"] == "PROJ"
-        elif "space" in simplified:
-            # The space field might be a dictionary with key and name fields
-            if isinstance(simplified["space"], dict):
-                assert simplified["space"]["key"] == "PROJ"
-                assert simplified["space"]["name"] == "Project Space"
-            # Or it might be a string with just the key
-            else:
-                assert (
-                    simplified["space"] == "PROJ"
-                    or simplified["space"] == "Project Space"
-                )
-
-        # Check version is included
-        assert "version" in simplified
-        assert simplified["version"] == 1
-
-        # URL should be included
-        assert "url" in simplified
-
-    def test_from_api_response_with_expandable_space(self):
-        """Test creating a ConfluencePage from data with space info in _expandable."""
-        page_data = {
-            "id": "123456",
-            "title": "Test Page",
-            "_expandable": {"space": "/rest/api/space/TEST"},
-        }
-
-        page = ConfluencePage.from_api_response(
-            page_data, base_url="https://confluence.example.com", is_cloud=True
-        )
-
-        assert page.space is not None
-        assert page.space.key == "TEST"
-        assert page.space.name == "Space TEST"
-        assert page.url == "https://confluence.example.com/spaces/TEST/pages/123456"
-
-    def test_from_api_response_with_missing_space(self):
-        """Test creating a ConfluencePage with no space information."""
-        page_data = {"id": "123456", "title": "Test Page"}
-
-        page = ConfluencePage.from_api_response(
-            page_data, base_url="https://confluence.example.com", is_cloud=True
-        )
-
-        assert page.space is not None
-        assert page.space.key == ""  # Default from ConfluenceSpace
-        assert page.url == "https://confluence.example.com/spaces/unknown/pages/123456"
-
-    def test_from_api_response_with_empty_space_data(self):
-        """Test creating a ConfluencePage with empty space data."""
-        page_data = {
-            "id": "123456",
-            "title": "Test Page",
-            "space": {},  # Empty space data
-        }
-
-        page = ConfluencePage.from_api_response(
-            page_data, base_url="https://confluence.example.com", is_cloud=True
-        )
-
-        assert page.space is not None
-        assert page.space.key == ""  # Default from ConfluenceSpace
-        assert page.url == "https://confluence.example.com/spaces/unknown/pages/123456"
-
-    def test_from_api_response_url_construction_without_base_url(self):
-        """Test that URL is None when base_url is not provided."""
-        page_data = {
-            "id": "123456",
-            "title": "Test Page",
-            "space": {"key": "TEST", "name": "Test Space"},
-        }
-
-        page = ConfluencePage.from_api_response(page_data)  # No base_url provided
-
-        assert page.url is None
-        assert page.space is not None
-        assert page.space.key == "TEST"
-
-    def test_url_construction_cloud_format(self):
-        """Test URL construction in cloud format."""
-        page_data = {
-            "id": "123456",
-            "title": "Test Page",
-            "space": {"key": "TEST", "name": "Test Space"},
-        }
-
-        page = ConfluencePage.from_api_response(
-            page_data, base_url="https://example.atlassian.net/wiki", is_cloud=True
-        )
-
-        assert page.url == "https://example.atlassian.net/wiki/spaces/TEST/pages/123456"
-
-    def test_url_construction_server_format(self):
-        """Test URL construction in server format."""
-        page_data = {
-            "id": "123456",
-            "title": "Test Page",
-            "space": {"key": "TEST", "name": "Test Space"},
-        }
-
-        page = ConfluencePage.from_api_response(
-            page_data, base_url="https://wiki.corp.example.com", is_cloud=False
-        )
-
-        assert (
-            page.url
-            == "https://wiki.corp.example.com/pages/viewpage.action?pageId=123456"
-        )
-
-    def test_attachment_url_uses_parent_page_id_server(self):
-        """Test that attachment URLs use parent page ID for server format."""
-        attachment_data = {
-            "id": "att105348",
-            "type": "attachment",
-            "title": "document.pdf",
-            "container": {"id": "12345", "type": "page"},
-        }
-
-        page = ConfluencePage.from_api_response(
-            attachment_data,
-            base_url="http://wiki.example.com",
-            is_cloud=False,
-        )
-
-        assert "pageId=12345" in page.url
-        assert "att105348" not in page.url
-
-    def test_attachment_url_uses_parent_page_id_cloud(self):
-        """Test that attachment URLs use parent page ID for cloud format."""
-        attachment_data = {
-            "id": "att105348",
-            "type": "attachment",
-            "title": "document.pdf",
-            "space": {"key": "TEST"},
-            "container": {"id": "12345", "type": "page"},
-        }
-
-        page = ConfluencePage.from_api_response(
-            attachment_data,
-            base_url="https://example.atlassian.net/wiki",
-            is_cloud=True,
-        )
-
-        assert "/pages/12345" in page.url
-        assert "att105348" not in page.url
-
-    def test_attachment_without_container_falls_back_to_own_id(self):
-        """Test that attachments without container fall back to their own ID."""
-        attachment_data = {
-            "id": "att105348",
-            "type": "attachment",
-            "title": "document.pdf",
-            # No container field
-        }
-
-        page = ConfluencePage.from_api_response(
-            attachment_data,
-            base_url="http://wiki.example.com",
-            is_cloud=False,
-        )
-
-        # Falls back to attachment ID when no container
-        assert "pageId=att105348" in page.url
-
-
-class TestConfluenceSearchResult:
-    """Tests for the ConfluenceSearchResult model."""
-
-    def test_from_api_response_with_valid_data(self, confluence_search_data):
-        """Test creating a ConfluenceSearchResult from valid API data."""
-        search_result = ConfluenceSearchResult.from_api_response(confluence_search_data)
-
-        assert search_result.total_size == 1
-        assert search_result.start == 0
-        assert search_result.limit == 50
-        assert search_result.cql_query == "parent = 123456789"
-        assert search_result.search_duration == 156
-
-        assert len(search_result.results) == 1
-
-        # Verify that results are properly converted to ConfluencePage objects
-        page = search_result.results[0]
-        assert isinstance(page, ConfluencePage)
-        assert page.id == "123456789"
-        assert page.title == "2024-01-01: Team Progress Meeting 01"
-
-    def test_from_api_response_with_empty_data(self):
-        """Test creating a ConfluenceSearchResult from empty data."""
-        search_result = ConfluenceSearchResult.from_api_response({})
-
-        # Should use default values
-        assert search_result.total_size == 0
-        assert search_result.start == 0
-        assert search_result.limit == 0
-        assert search_result.cql_query is None
-        assert search_result.search_duration is None
-        assert len(search_result.results) == 0
-
-
-class TestRealConfluenceData:
-    """Tests using real Confluence data (only run if environment is configured)."""
-
-    def test_real_confluence_page(
-        self, use_real_confluence_data, default_confluence_page_id
-    ):
-        """Test with real Confluence page data from the API."""
-        if not use_real_confluence_data:
-            pytest.skip("Real Confluence data testing is disabled")
-
-        try:
-            # Initialize the Confluence client
-            from src.mcp_atlassian.confluence.client import ConfluenceClient
-            from src.mcp_atlassian.confluence.config import ConfluenceConfig
-            from src.mcp_atlassian.confluence.pages import PagesMixin
-
-            # Use the from_env method to create the config
-            config = ConfluenceConfig.from_env()
-            confluence_client = ConfluenceClient(config=config)
-            pages_client = PagesMixin(config=config)
-
-            # Use the provided page ID from environment or fixture
-            page_id = default_confluence_page_id
-
-            # Get page data directly from the Confluence API
-            page_data = confluence_client.confluence.get_page_by_id(
-                page_id=page_id, expand="body.storage,version,space,children.attachment"
-            )
-
-            # Convert to model
-            from src.mcp_atlassian.models import ConfluencePage
-
-            page = ConfluencePage.from_api_response(page_data)
-
-            # Verify basic properties
-            assert page.id == page_id
-            assert page.title is not None
-            assert page.space is not None
-            assert page.space.key is not None
-            assert page.attachments is not None
-
-            # Verify that to_simplified_dict works
-            simplified = page.to_simplified_dict()
-            assert isinstance(simplified, dict)
-            assert simplified["id"] == page_id
-
-            # Get and test comments if available
-            try:
-                from src.mcp_atlassian.models import ConfluenceComment
-
-                comments_data = confluence_client.confluence.get_page_comments(
-                    page_id=page_id, expand="body.view,version"
-                )
-
-                if comments_data and comments_data.get("results"):
-                    comment_data = comments_data["results"][0]
-                    comment = ConfluenceComment.from_api_response(comment_data)
-
-                    assert comment.id is not None
-                    assert comment.body is not None
-
-                    # Test simplified dict
-                    comment_dict = comment.to_simplified_dict()
-                    assert isinstance(comment_dict, dict)
-                    assert "body" in comment_dict
-            except Exception as e:
-                print(f"Comments test skipped: {e}")
-
-            print(
-                f"Successfully tested real Confluence page {page_id} in space {page.space.key}"
-            )
-        except ImportError as e:
-            pytest.skip(f"Could not import Confluence client: {e}")
-        except Exception as e:
-            pytest.fail(f"Error testing real Confluence page: {e}")
